move excel functions to separate file
