+++ backend/web.py
@@ -16,6 +16,7 @@ import csv
 import dictionaries
 from electionRules import ElectionRules
 import util
+import excel_util
 import voting
 import simulate as sim
 
@@ -90,7 +91,7 @@ def get_election_excel():
         return jsonify(election)
 
     tmpfilename = tempfile.mktemp(prefix='election-')
-    util.election_to_xlsx(election, tmpfilename)
+    excel_util.election_to_xlsx(election, tmpfilename)
     print("%s" % (tmpfilename))
     return send_from_directory(
         directory=os.path.dirname(tmpfilename),
@@ -228,7 +229,7 @@ def get_xlsx():
 
     tmpfilename = tempfile.mktemp(prefix='votesim-%s-' % request.args["sid"][:6])
     simulation, thread, expiry = SIMULATIONS[request.args["sid"]]
-    util.simulation_to_xlsx(simulation, tmpfilename)
+    excel_util.simulation_to_xlsx(simulation, tmpfilename)
     print("%s" % (tmpfilename))
     return send_from_directory(
         directory=os.path.dirname(tmpfilename),
